requirements.txt
annotated-types==0.7.0
anyio==4.9.0
fastapi==0.115.14
idna==3.10
pydantic==2.11.7
pydantic_core==2.33.2
python-dotenv==1.1.1
sniffio==1.3.1
starlette==0.46.2
typing-inspection==0.4.1
typing_extensions==4.14.0
